Move activerecord model into fixtures directory
